transport/tchannel: move handler into separate file
Add inboundCall interface to make testing the handler easier.

Delete unnecessary Inbound interface.
