+++ transport/tchannel/inbound.go
@@ -22,24 +22,11 @@ package tchannel
 
 import (
 	"net"
-	"time"
 
 	"github.com/yarpc/yarpc-go/transport"
 
 	"github.com/uber/tchannel-go"
-	"golang.org/x/net/context"
 )
-
-// Inbound represents a TChannel Inbound. It is the same as a transport
-// Inbound except it exposes the address on which the system is listening for
-// connections.
-type Inbound interface {
-	transport.Inbound
-
-	// Address on which the server is listening. Returns nil if Start has not
-	// been called yet.
-	Addr() net.Addr
-}
 
 // InboundOption configures Inbound.
 type InboundOption func(*inbound)
@@ -54,7 +41,7 @@ func ListenAddr(addr string) InboundOption {
 }
 
 // NewInbound builds a new TChannel inbound from the given Channel.
-func NewInbound(ch *tchannel.Channel, opts ...InboundOption) Inbound {
+func NewInbound(ch *tchannel.Channel, opts ...InboundOption) transport.Inbound {
 	i := &inbound{ch: ch}
 	for _, opt := range opts {
 		opt(i)
@@ -109,117 +96,5 @@ func (i *inbound) Start(h transport.Handler) error {
 
 func (i *inbound) Stop() error {
 	i.ch.Close()
-	return nil
-}
-
-func (i *inbound) Addr() net.Addr {
-	if i.listener == nil {
-		return nil
-	}
-	return i.listener.Addr()
-}
-
-type handler struct{ Handler transport.Handler }
-
-func (h handler) Handle(ctx context.Context, call *tchannel.InboundCall) {
-	deadline, ok := ctx.Deadline()
-	if !ok {
-		call.Response().SendSystemError(tchannel.ErrTimeoutRequired)
-		return
-	}
-
-	headers, err := readHeaders(call.Format(), call.Arg2Reader)
-	if err != nil {
-		call.Response().SendSystemError(tchannel.NewSystemError(
-			tchannel.ErrCodeUnexpected, "failed to read headers: %v", err))
-		return
-	}
-
-	body, err := call.Arg3Reader()
-	if err != nil {
-		call.Response().SendSystemError(tchannel.NewSystemError(
-			tchannel.ErrCodeUnexpected, "failed to read body: %v", err))
-		return
-	}
-	defer body.Close()
-
-	rw := newResponseWriter(call)
-	defer rw.Close() // TODO(abg): log if this errors
-
-	treq := &transport.Request{
-		Caller:    call.CallerName(),
-		Service:   call.ServiceName(),
-		Encoding:  transport.Encoding(call.Format()),
-		Procedure: call.MethodString(),
-		Headers:   headers,
-		Body:      body,
-		TTL:       deadline.Sub(time.Now()),
-	}
-
-	if err := h.Handler.Handle(ctx, treq, rw); err != nil {
-		call.Response().SendSystemError(tchannel.NewSystemError(
-			tchannel.ErrCodeUnexpected, "internal error: %v", err))
-		return
-	}
-}
-
-type responseWriter struct {
-	failedWith   error
-	bodyWriter   tchannel.ArgWriter
-	format       tchannel.Format
-	headers      transport.Headers
-	response     *tchannel.InboundCallResponse
-	wroteHeaders bool
-}
-
-func newResponseWriter(call *tchannel.InboundCall) *responseWriter {
-	return &responseWriter{
-		response: call.Response(),
-		headers:  make(transport.Headers),
-		format:   call.Format(),
-	}
-}
-
-func (rw *responseWriter) AddHeaders(h transport.Headers) {
-	if rw.wroteHeaders {
-		panic("AddHeaders() cannot be called after calling Write().")
-	}
-	for k, v := range h {
-		rw.headers.Set(k, v)
-	}
-}
-
-func (rw *responseWriter) Write(s []byte) (int, error) {
-	if rw.failedWith != nil {
-		return 0, rw.failedWith
-	}
-
-	if !rw.wroteHeaders {
-		rw.wroteHeaders = true
-		if err := writeHeaders(rw.format, rw.headers, rw.response.Arg2Writer); err != nil {
-			rw.failedWith = err
-			return 0, err
-		}
-	}
-
-	if rw.bodyWriter == nil {
-		var err error
-		rw.bodyWriter, err = rw.response.Arg3Writer()
-		if err != nil {
-			rw.failedWith = err
-			return 0, err
-		}
-	}
-
-	return rw.bodyWriter.Write(s)
-}
-
-func (rw *responseWriter) Close() error {
-	if rw.bodyWriter != nil {
-		return rw.bodyWriter.Close()
-	}
-	if rw.failedWith != nil {
-		return rw.failedWith
-	}
 	return nil
 }
